+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 V2.1.1 (2020-02-09)
 * Limit the amount of time the system may be in the Wake mode to the long poll time
 * Fixed optional rear seat heater handling
+* Added individual security overrides for lock, sunroof, charge_port, windows, trunk, frunk, sentry, and software_update
 
 v2.1.0 (2020-01-18)
 * When the General node is in wake mode, all three nodes will be refreshed including the Climate and Security nodes